Rework Gitlab CI to build CI images automatically on schedule
With this change, CI images are now built in the CI itself, so that it
can be automatically rebuilt on schedule, say every week or so.

Additionally, all tests that were marked `only: [tags]` are now marked
`only: [schedules, tags]`, meaning that if I schedule the pipeline
weekly, every week I'll get 1) Rebuild of CI images and 2) Run all the
tests with these new images.