Fix how compile-time version strings are displayed
Before:

  Compiled against: GLib 2.74.6, Soup (3).(2).(2), GStreamer (1).(22).(0).(0), GTK (3).(24).(36)
  Running along   : GLib 2.74.6, Soup 3.2.2, GStreamer 1.22.0.0, GTK 3.24.36

Most libraries actually surround their version with parentheses, and
those parentheses are converted to string by G_STRINGIFY(). Cf.
https://gitlab.gnome.org/GNOME/libsoup/-/issues/339

So we need an extra macro for those.

Enters:

  #define GV_STRINGIFY_DEPAREN(x) G_STRINGIFY_ARG x

To be used in place of:

  #define G_STRINGIFY(macro_or_string) G_STRINGIFY_ARG (macro_or_string)

Result:

  Compiled against: GLib 2.74.6, Soup 3.2.2, GStreamer 1.22.0.0, GTK 3.24.36
  Running along   : GLib 2.74.6, Soup 3.2.2, GStreamer 1.22.0.0, GTK 3.24.36

Note that I we don't define GV_STRINGIFY_DEPAREN in ui/gv-ui.h, nor do
we include core/gv-core.h, and yet it compiles, so be it.