Switch to libsoup3
Update gv-engine and gv-station to use libsoup3.

Let's look at gv-engine first: with this commit, we lost error handling
in case of SSL errors...  It's lost for good, as we just can't do it as
before, but I started a discussion at:
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/2379

So what's different with libsoup3? When the certificate is not valid, it
results in a GST_STREAM_ERROR_FAILED error. The error message is
Internal data stream error. If I do gst_message_parse_error_details, all
I get is a field flow-return=(int)-5.

So there is absolutely no way to know that it's a SSL error.

The field http-status-code that I relied on is gone. In any case, this
field contained a "libsoup custom code", not a real HTTP status code, so
it was an oddity, and it's gone for good reasons.

Regarding gv-station now: so far I didn't handle SSL errors, for
playlist download. I never implemented it, in part because I knew that
it was changing in libsoup3. Good news is that, now that I'm switching
to libsoup3, I'll add support for these errors in the next commits.