station: Add a redirected-uri property
This is mostly to be able to show the real URL, in case of a SSL
failure. It happens that a http:// link redirects to https://, and then
the certificate is invalid. In such case, showing "certificate error for
http://..." doesn't make any sense.

There's an issue though: it's yet another state within a GvStation
object (after stream-uris), and for now this state is "forever". But a
redirection can be temporary or something, so every time we play a
station we should clear this state, and restart from zero.

This will be addressed in upcoming commits.