core: gv_engine_play takes a uri, don't deal with GvStation anymore
Passing a GvStation to gv_engine_play() has always been awkward, and now
that we have gv_station_play() calling gv_engine_play(), it becomes even
more clear that GvEngine should know about GvStation, but it should
rather just deal with a URI.

Of course we still need to know about the user-agent (which can be per
station), and also ssl-strict (which is broken at the moment). For now I
just add those as params to gv_engine_play().