ui: Add an 'anchor' property to GvStationDialog
We're going to drop the last gtk_dialog_run(), meaning we go fully async
with dialogs, hence we need a place to store the station that was
righy-clicked and serves as an anchor. Having it in the context menu
object is not enough, as the context menu vanishes straight after the
dialog is opened (the dialog's parent is the main window, not the
context menu).

So let's save the anchor in the GvStationDialog object itself, so that
we can retrieve it later.