Rename signal ssl-failure to bad-certificate, drop uri arg
The name ssl came from GStreamer (and its ssl-strict property), however
nowadays SSL is replaced by TLS (and it's the name used in GLib, eg.
GTls...). So I was never quite happy with the term 'ssl-failure'.

Finally, I think 'bad-certificate' is a better fit, it's future-proof
and it's in line with other signal's name like 'accept-certificate'.

Also (and maybe the most important), this commit drops the 'uri'
argument from the signal, it doesn't really cut it, it will work better
to just let the client get the uri value by itself, and more important,
update it if ever the uri change afterwards. Reason is: we might receive
the signal 'accept-certificate' before being notified of a redirection.