ui: Add new widget GvVolumeControl
The volume control widget is simply a horizontal box with two elements,
a mute button and a volume slider. It has the same look and feel as the
volume control that you can find in the Youtube player, and in fact in
most web players out there. The combination of a mute icon + a volume
slider is pretty ubiquitous on the web.