volume-control: Initialize the fallback volume on map signal
When Goodvibes is started, usually volume is 100% and mute is false. It
matches the defaults in GvPlayer, hence no signals are emitted, and no
callbacks get called in GvVolumeControl, so the fallback volume is
unset.

If users drags the volume to zero, then clicks unmute, what happens is
that:
- current volume is zero (since user dragged it there)
- no fallback volume is set (since volume was never changed)

In this case, volume is restored to 50%: not what we want.

So we just need to make sure to initialize the fallback volume when the
widget is mapped, which should have been done in
2d1bfa290eeec36d019d75ca0412948d152809b7.