Update gstreamer dependency to 1.24
Due to switch to libsoup 3.x, and the fact that Goodvibes tries hard to
catch SSL errors and report it to users, we need 1.24 for this
functionality to work. Without 1.24, SSL errors are simply ignored by
Goodvibes.

Another thing is the switch to playbin3, which was marked experimental
for a long time, and are considered stable only from GStreamer 1.24 [1].

Finally I don't have time to test older versions.

So let's just move on and make 1.24 the new minimum requirement.

[1]: https://gstreamer.freedesktop.org/releases/1.24/